backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.27.0
anthropic>=0.40.0
supabase>=2.3.0
apscheduler>=3.10.0
pydantic>=2.6.0
feedparser>=6.0.0
yfinance>=0.2.36
firebase-admin>=6.4.0
slowapi>=0.1.9
bcrypt>=4.1.0
PyJWT>=2.8.0
pyotp>=2.9.0
